CHA_0035_CreateGtosMto

+++ Condominios/Condominios/Data/Interfaces/IEpoch.cs
@@ -3,7 +3,7 @@
     public interface IEpoch
     {
         public long CrearEpoch(DateTime fecha);
-        public DateTime ObtenerFecha(double epoch);
+        public DateTime ObtenerFecha(long epoch);
         public string ObtenerMesYAnio(DateTime fecha);
     }
 }
